+++ qiskit_ibm_runtime/results/__init__.py
@@ -30,7 +30,7 @@ Result classes (:mod:`qiskit_ibm_runtime.results`)
    QuantumProgramItemResult
 """  # noqa: D205, D212, D415
 
-from .estimator import EstimatorPubResult
+from .estimator_pub import EstimatorPubResult
 from .noise_learner import LayerError, NoiseLearnerResult, PauliLindbladError
 from .noise_learner_v3 import NoiseLearnerV3Result, NoiseLearnerV3Results
 from .quantum_program import ItemMetadata, QuantumProgramItemResult, QuantumProgramResult
